Button -> hoverEvent(), clickEvent()

+++ src/main/java/org/collebol/client/gui/graphics/renderer/ui/TextInputRenderer.java
@@ -28,7 +28,7 @@ public class TextInputRenderer extends Renderer {
      */
     public void renderSubTextInput(int id, int parentId) {
         Field parent = (Field) this.engine.getComponentHandler().getComponent(Field.class, parentId);
-        renderTextInputInternal(id, (float) parent.getPosition().getX(), (float) parent.getPosition().getY());
+        renderTextInputInternal(id, parent.getPosition().getX(), parent.getPosition().getY());
     }
 
     /**
